Starts implementation of 7-3

+++ crates/cibse_guide_e/src/chapter_7/equation_7_2.rs
@@ -1,7 +1,7 @@
 pub mod openfire_runner;
 
-pub fn stair_capacity(w: f64, n: f64) -> f64 {
-    let result: f64 = 200.0 * w + 50.0 * (w - 0.3) * (n - 1.0);
+pub fn stair_capacity(w: f64, n: i16) -> f64 {
+    let result: f64 = 200.0 * w + 50.0 * (w - 0.3) * (n - 1) as f64;
     return result.floor();
 }
 
@@ -18,7 +18,7 @@ mod tests {
 
     #[test]
     fn test() {
-        let result = stair_capacity(1.2, 6.0);
+        let result = stair_capacity(1.2, 6);
         assert_eq!(result, 465.0);
     }
 }
